# Improve RSA/ECDSA performance